fixed gdscript-float -> sfloat conversion fixed unconsistent method calling pattern

+++ scripts/network/SyncManagerWrapper.cs
@@ -13,7 +13,7 @@ namespace Bonebreaker.Network
             _SM = GetTree().Root.GetNode("SyncManager");
         }
 
-        public static sfloat GetDeltaTime () => (sfloat)_instance._SM.Get("tick_time");
+        public static sfloat GetDeltaTime () => (sfloat)(float)_instance._SM.Get("tick_time");
         public static int GetCurrentTick () => (int)_instance._SM.Get("current_tick");
     }
 }
